requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Limiter==3.5.0
Flask-WTF==1.1.1
pypdf==3.17.4
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
gunicorn==21.2.0
python-dotenv==1.0.0
